requirements.txt
aiogram
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
python-dotenv>=1.0.0